requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.6
django-celery-beat==2.5.0
pandas==2.1.4
plotly==5.18.0
django-plotly-dash==2.2.2
python-dateutil==2.8.2